Add weighted random
This is just a test commit to check if the weighted colour generation is
even worth adding.

Note: this will not work if the consecutive monocolour option is set,
because it might be too repetitive, favouring the monocolor patterns,
e.g. when the option is set to 4, this just might create `ddddk(...)`
too many times.